--- escli_tool/data/processor_base.py
@@ -1,6 +1,0 @@
-
-
-
-class ProcessorBase:
-    def __init__(self):
-        pass
