+++ escli_tool/data_processor/processor_base.py
@@ -1,0 +1,10 @@
+from abc import ABC, abstractmethod
+
+
+class ProcessorBase(ABC):
+    def __init__(self, commit_id: str, commit_title, created_at: str):
+        self.commit_id = commit_id
+        self.commit_title = commit_title
+        self.created_at = created_at
+    
+
