+++ escli_tool/handler.py
@@ -6,7 +6,7 @@ from typing import List
 import requests
 import urllib3
 
-from escli_tool.data import common
+from escli_tool import common
 from escli_tool.utils import get_logger, load_credentials, read_from_json
 
 urllib3.disable_warnings(urllib3.exceptions.InsecureRequestWarning)
